+++ doc/release_notes.rst
@@ -7,8 +7,16 @@ desimodel Release Notes
 
 * Add new weather module to specify assumed atmospheric seeing and transparency
   distributions at KPNO, with accompanying DESI-doc and jupyter notebook.
-* Remove seeing module, which is superceded by new weather module.
+* Remove seeing module, which is superseded by new weather module.
+* Added `desimodel.footprint.pixweight()` in :mod:`desimodel.footprint` to create an array of what fraction
+  of every HEALPixel at a given nside overlaps the DESI footprint
+* Also added `desimodel.footprint.tiles2fracpix()` to estimate which HEALPixels overlap the footprint edges
+* Added `desimodel.io.load_pixweight()` in :mod:`desimodel.io` to load the array created by
+  `desimodel.footprint.pixweight()` and resample it to any HEALPix nside
 * Modified path to Lya SNR spectra files used in desi_quicklya.py, used in Lya Fisher forecast.
+* Added desimodel.inputs.build_gfa_table and its helper functions to write a .ecsv file for GFA data
+* Added desimodel.io.load_gfa to return the GFA data table
+* Added desimodel.focalplane.xy2radec, which converts x,y coordinates on the focal plane to RA, Dec coordinates
 * don't print warnings in desimodel.io if specter isn't installed
 
 0.7.0 (2017-06-15)